+++ src/components/games/VocabularyGames.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { motion } from 'framer-motion';
-import { Card, CardContent, CardHeader, CardTitle } from './ui/card';
-import { Button } from './ui/button';
-import { Badge } from './ui/badge';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
+import { Button } from '../ui/button';
+import { Badge } from '../ui/badge';
 import MatchingCardsGame from './MatchingCardsGame';
 import WordAssociationGame from './WordAssociationGame';
 import FlashcardMemoryGame from './FlashcardMemoryGame';
