+++ src/components/games/WordAssociationGame.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
-import { Card, CardContent, CardHeader, CardTitle } from './ui/card';
-import { Button } from './ui/button';
-import { Badge } from './ui/badge';
-import { Progress } from './ui/progress';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
+import { Button } from '../ui/button';
+import { Badge } from '../ui/badge';
+import { Progress } from '../ui/progress';
 import SetupScreen from './word-association/SetupScreen';
 import CompletionScreen from './word-association/CompletionScreen';
 import PauseOverlay from './word-association/PauseOverlay';
