+++ src/components/games/word-association/CompletionScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { motion } from 'framer-motion';
-import { Card, CardContent } from '../ui/card';
-import { Button } from '../ui/button';
+import { Card, CardContent } from '../../ui/card';
+import { Button } from '../../ui/button';
 import { RotateCcw, Home, Trophy } from 'lucide-react';
 
 interface CompletionScreenProps {
