+++ src/components/games/word-association/modes/synAnt.ts
