+++ src/components/home/Categories.tsx
