+++ src/components/home/Statistics.tsx
