+++ src/components/quiz/QuestionTemplates.tsx
